refactor: based on comments in #21
