+++ src/lib.rs
@@ -1,0 +1,4 @@
+//! The library backend for the Bevy CLI.
+
+pub mod lint;
+pub mod template;
